+++ check-in-request-MS/src/main/java/com/example/checkinrequestMS/CheckInRequestMsApplication.java
@@ -1,4 +1,4 @@
-package com.example.checkinrequestms;
+package com.example.checkinrequestMS;
 
 
 import org.springframework.boot.SpringApplication;
