+++ check-in-request-MS/src/main/resources/application.yml
@@ -1,0 +1,2 @@
+kakao:
+  apikey: a0dc4e7625b15b5b4cef4e0a028119b3
