+++ check-in-request-MS/src/test/java/com/example/checkinrequestMS/CheckInRequestMsApplicationTests.java
@@ -1,4 +1,4 @@
-package com.example.checkinrequestms;
+package com.example.checkinrequestMS;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
