--- check-in-request-MS/src/test/java/com/example/checkinrequestms/HelpAPI/web/rest/KakaoAPIRequestTest.java
@@ -1,13 +1,0 @@
-package com.example.checkinrequestms.HelpAPI.web.rest;
-
-import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class KakaoAPIRequestTest {
-
-    @Test
-    void test1() {
-        KakaoAPIRequest.test();
-    }
-}
